% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_earliest_date.R
\name{earliest_date}
\alias{earliest_date}
\title{ealiest_date}
\usage{
earliest_date(x, start_date = start_date, end_date = end_date, ...)
}
\arguments{
\item{...}{Numeric, complex, or logical vectors.
be removed?}
}
\value{
If all inputs are integer and logical, then the output
}
\description{
\code{get_devices} Gets device numbers from GAP API.
}
\details{
This is a wrapper function for \code{httr::GET} which makes use of
  \code{format_uri} to formulate the correct uri string, and
  \code{jsonlite::fromJSON} to parse the incoming json, and output a
  character vector.
}
\examples{
TODO
}

