% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_format_uri.R
\name{format_uri}
\alias{format_uri}
\title{int_format_url}
\usage{
format_uri(tenant = tenant, device = device, type = "messages",
  filter_string = "")
}
\arguments{
\item{tenant}{The tenant code required for access to the GAP API. This is provided by CoSignal support.}

\item{device}{device number retried from \code{get_devices}.}

\item{type}{character string either: \code{messages} or \code{devices}}
}
\value{
If all inputs are integer and logical, then the output
  will be an integer. If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{int_format_url} Generic internal function that formats api url string.
}
\details{
Generic internal function that formats a url string for the gap API.
}
\examples{
sum("a")
}

