% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_devices.R
\name{get_devices}
\alias{get_devices}
\title{Get device numbers}
\usage{
get_devices(tenant, username, password, uri_string = "")
}
\arguments{
\item{tenant}{The tenant code required for access to the GAP API. This is
provided by CoSignal support.}

\item{username}{Username used to access the ninja tracking control panel.}

\item{password}{Password used to access the ninjac tracking control panel.
removed?}
}
\value{
A character vector of device IDs for the appropriate tenant.
}
\description{
\code{get_devices} Gets device numbers from GAP API.
}
\details{
This is a wrapper function for \code{httr::GET} which makes use of
  \code{format_uri} to formulate the correct uri string, and
  \code{jsonlite::fromJSON} to parse the incoming json, and output a
  character vector.
}
\examples{
library(dplyr)

tenant <- <tenant>
username <- <username>
password <- <password>

device_list <- get_devices(
tenant, username, password
)

df <- get_messages(
device = device_list[1],
tenant = tenant,
username = username,
password = password
) \%>\%
simplify_cattle(
device = device_list[1]
)

print(df)
}

