% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_GPS.R
\name{get_GPS}
\alias{get_GPS}
\title{get_GPS.}
\usage{
get_GPS(tenant, username, password, device_list,
  start_date = "1900-01-01T00:00:00Z", end_date = NULL)
}
\arguments{
\item{tenant}{The tenant code required for access to the GAP API. This is
provided by CoSignal support.}

\item{username}{Username used to access the ninja tracking control panel.}

\item{password}{Password used to access the ninjac tracking control panel.
removed?}

\item{device_list}{A character vector of device ids over which the function
will iterate. These are obtained with \code{get_devices}.}

\item{start_date}{date from which observations will begin, given as a
character vector of length one in the format: yyyy-mm-ddT00:00:00Z.
Defaults to 1900-01-01T00:00:00Z. Currently functionality to select by date
has not been implemented, but will be in a future version.}

\item{end_date}{date at which observations will end, given as a character
  vector of length one in the format: yyyy-mm-ddT00:00:00Z. Defaults to null.
  Currently functionality to select by date has not been implemented, but
  will be in a future version}
}
\value{
A character vector of device IDs for the appropriate tenant.
}
\description{
\code{get_GPS} get GPS measurements from devices.
}
\details{
This function accesses the GAP API to extract GPS meassages for a
  given device number between a date range. The function is essentially a
  wrapper for the \code{gapAPI::get_messages} and internal
  \code{gapAPI::simplify_format} functions.
}
\examples{
messages <- get_GPS <- function(
device = <device>,
tenant = <tenant-code>,
username = <username>,
password = <password>
)
print(messages)
}

