% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_records.R
\name{save_records}
\alias{save_records}
\title{Save GPS data to .csv}
\usage{
save_records(x, folder)
}
\arguments{
\item{x}{a dataframe object created by \code{simplify_cattle}. Can also be
used for other dataframes with a SentFromDevice field.}

\item{folder}{destination folder.}
}
\value{
Writes out csv file to destination folder with file name of the format: <device>_yyyy-mm-dd_HH_MM_SS_.csv
}
\description{
Saves objects created by \code{get_messages} and
  \code{simplify_cattle} to a .csv file.
}
\details{
Saves out .csv file from GPS data obtained using \code{get_messages}
  and then passed to \code{simplify_cattle}, and includes a check that an
  output file was created.
}
\examples{
library(dplyr)

tenant <- <tenant>
username <- <username>
password <- <password>

device_list <- get_devices(
tenant, username, password
)

# Must create "my folder" first

df <- get_messages(
device = device_list[1],
tenant = tenant,
username = username,
password = password
) \%>\%
simplify_cattle(
device = device_list[1]
) \%>\%
save_records(
"my_folder"
)

print(df)
}

