% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int_simplify_cattle.R
\name{simplify_cattle}
\alias{simplify_cattle}
\title{Simplify output from \code{get_messages}}
\usage{
simplify_cattle(x, device = device)
}
\arguments{
\item{x}{an object returned by \code{get_messages}}

\item{device}{the device from which the \code{get_messages} object was
  obtained. Future versions will do this automatically.}
}
\value{
An object of class \code{c(tbl,tbl_df),data.frame} with the following
  columns: device, timestamp, date, time, SequenceNumber,SentFromDevice,
  MessageType, lon, lat, speed, heading, alt, accuracy.
}
\description{
\code{simplify_cattle} data manipulaton function which formats outputs from
\code{get_messages}.
}
\details{
This function implements some simple data manipulation operations for objects
returned by \code{get_messages}
}
\examples{
library(dplyr)

tenant <- <tenant>
username <- <username>
password <- <password>

device_list <- get_devices(
tenant, username, password
)

df <- get_messages(
device = device_list[1],
tenant = tenant,
username = username,
password = password
) \%>\%
simplify_cattle(
device = device_list[1]
)

print(df)
}

